+++ example/integration_test/screens/image_settings.dart
@@ -3,7 +3,7 @@
  * See LICENSE for distribution and usage details.
  */
 
-import '../utils/export.dart';
+import '../utils/image_links.dart';
 
 import 'package:example/main.dart';
 
@@ -55,6 +55,7 @@ void testSuite({
         type: l10n.isBackground,
         updateCS: true,
         l10n: l10n,
+        networkImageURLs: imageURLs,
         isLefty: isLefty,
       );
 
@@ -69,6 +70,6 @@ void testSuite({
 
       // Reset for next test suite  //
 
-      await goBack(tester, l10n);
+      await goBack(tester, l10n.gBack);
       debugPrint('\nImage settings test suite complete\n\n');
     });