+++ lib/empathetech_flutter_ui.dart
@@ -19,6 +19,7 @@ export 'src/consts/google_fonts.dart';
 
 // Functions //
 
+// General
 export 'src/functions/colors.dart';
 export 'src/functions/dialogs.dart';
 export 'src/functions/feedback_localizations.dart';
@@ -26,6 +27,15 @@ export 'src/functions/helpers.dart';
 export 'src/functions/images.dart';
 export 'src/functions/text.dart';
 export 'src/functions/theme_data.dart';
+
+// Testing
+export 'src/functions/testing/consts.dart';
+export 'src/functions/testing/utils.dart';
+export 'src/functions/testing/widgets/dominant_hand_switch.dart';
+export 'src/functions/testing/widgets/image_setting.dart';
+export 'src/functions/testing/widgets/locale_setting.dart';
+export 'src/functions/testing/widgets/reset_button.dart';
+export 'src/functions/testing/widgets/theme_mode_switch.dart';
 
 // l10n //