+++ lib/src/functions/testing/widgets/reset_button.dart
@@ -1,9 +1,7 @@
-/* open_ui
+/* empathetech_flutter_ui
  * Copyright (c) 2022-2024 Empathetech LLC. All rights reserved.
  * See LICENSE for distribution and usage details.
  */
-
-import '../export.dart';
 
 import 'dart:io';
 import 'package:flutter/material.dart';
@@ -14,6 +12,8 @@ import 'package:empathetech_flutter_ui/empathetech_flutter_ui.dart';
 
 enum RBType { all, text, layout, color, image }
 
+/// For integration testing
+/// Tests [EzResetButton]
 Future<void> testResetButton(
   WidgetTester tester, {
   required RBType type,